Added hash bit avalanching for int and long keys with no comparers.
When int key is used as a hash, it is too easy to get into bad situations if keys are sparse. Masking of the hash can drop most of the hash entropy leading to collisions or clustering.
Avalanching of hash bits mitigates this possibility.

Overal this change trades perf of sequential dense scenarios (hashing on 1,2,3,4,5...) for improved perf of sparse random scenarios (hashing on 1, 1000, 17, 1024...)