Fix a race condition for Ref and Boxed keys between claiming and finding a slot that matches a given key.
Slots are claimed in two stages: claim a hash, then set a key
it is possible to observe a slot with a null key, but with hash already set

We did not check for the possibility of null entryKey, which could be unexpected by Ref comparer, or cause NRE in Boxed case when entryKey is unwrapped.

NOTE: this null-check is highly predictable since it nearly never sees actual nulls.
It does not appear to affect performance in any measurable way.